+++ geom/type_geometry.go
@@ -975,6 +975,35 @@ func (g Geometry) Validate() error {
 	}
 }
 
+// Densify returns a new Geometry with additional linearly interpolated control
+// points such that the distance between any two consecutive control points is
+// at most the given maxDistance.
+//
+// Panics if maxDistance is zero or negative.
+func (g Geometry) Densify(maxDistance float64) Geometry {
+	switch g.gtype {
+	case TypePoint, TypeMultiPoint:
+		// Points cannot be densified, but still check that the max distance is
+		// valid for consistency between types.
+		if maxDistance <= 0 {
+			panic("maxDistance must be positive")
+		}
+		return g
+	case TypeLineString:
+		return g.MustAsLineString().Densify(maxDistance).AsGeometry()
+	case TypeMultiLineString:
+		return g.MustAsMultiLineString().Densify(maxDistance).AsGeometry()
+	case TypePolygon:
+		return g.MustAsPolygon().Densify(maxDistance).AsGeometry()
+	case TypeMultiPolygon:
+		return g.MustAsMultiPolygon().Densify(maxDistance).AsGeometry()
+	case TypeGeometryCollection:
+		return g.MustAsGeometryCollection().Densify(maxDistance).AsGeometry()
+	default:
+		panic("unknown type: " + g.Type().String())
+	}
+}
+
 // SnapToGrid returns a copy of the geometry with all coordinates snapped to a
 // base 10 grid.
 //