+++ geom/type_line_string.go
@@ -474,6 +474,15 @@ func (s LineString) InterpolateEvenlySpacedPoints(n int) MultiPoint {
 	return NewMultiPoint(pts)
 }
 
+// Densify returns a new LineString with additional linearly interpolated
+// control points such that the distance between any two consecutive control
+// points is at most the given maxDistance.
+//
+// Panics if maxDistance is zero or negative.
+func (s LineString) Densify(minDistance float64) LineString {
+	return NewLineString(densify(s.seq, minDistance))
+}
+
 // SnapToGrid returns a copy of the LineString with all coordinates snapped to
 // a base 10 grid.
 //