+++ geom/type_multi_line_string.go
@@ -511,6 +511,19 @@ func (m MultiLineString) Simplify(threshold float64) MultiLineString {
 	return NewMultiLineString(lss)
 }
 
+// Densify returns a new MultiLineString with additional linearly interpolated
+// control points such that the distance between any two consecutive control
+// points is at most the given maxDistance.
+//
+// Panics if maxDistance is zero or negative.
+func (m MultiLineString) Densify(maxDistance float64) MultiLineString {
+	lss := make([]LineString, len(m.lines))
+	for i := range m.lines {
+		lss[i] = m.lines[i].Densify(maxDistance)
+	}
+	return MultiLineString{lss, m.ctype}
+}
+
 // SnapToGrid returns a copy of the MultiLineString with all coordinates
 // snapped to a base 10 grid.
 //