+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 YYYY-MM-DD
 
+- Adds new `Densify` methods on each concrete geometry types that add
+  additional linearly interpolated control points such that the distance
+  between any two consecutive control points is at most a fixed distance. This
+  can be useful when transforming geometries from one projection to another.
+
 - Adds a new method `SnapToGrid` to each concrete geometry type. The method
   returns a copy of the geometry with its control points snapped to a base 10
   grid.